requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==21.4
python-dotenv==1.0.1
